+++ stratosphere/src/services/aws/elasticsearch.rs
@@ -172,8 +172,8 @@ pub mod domain {
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticsearch-domain-ebsoptions.html
     pub struct EBSOptions_ {
         pub ebs_enabled: Option<crate::value::ExpBool>,
-        pub iops: Option<i64>,
-        pub volume_size: Option<i64>,
+        pub iops: Option<i32>,
+        pub volume_size: Option<i32>,
         pub volume_type: Option<crate::value::ExpString>,
     }
     #[doc(hidden)]
@@ -215,12 +215,12 @@ pub mod domain {
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticsearch-domain-elasticsearchclusterconfig.html
     pub struct ElasticsearchClusterConfig_ {
         pub cold_storage_options: Option<Box<ColdStorageOptions_>>,
-        pub dedicated_master_count: Option<i64>,
+        pub dedicated_master_count: Option<i32>,
         pub dedicated_master_enabled: Option<crate::value::ExpBool>,
         pub dedicated_master_type: Option<crate::value::ExpString>,
-        pub instance_count: Option<i64>,
+        pub instance_count: Option<i32>,
         pub instance_type: Option<crate::value::ExpString>,
-        pub warm_count: Option<i64>,
+        pub warm_count: Option<i32>,
         pub warm_enabled: Option<crate::value::ExpBool>,
         pub warm_type: Option<crate::value::ExpString>,
         pub zone_awareness_config: Option<Box<ZoneAwarenessConfig_>>,
@@ -437,7 +437,7 @@ pub mod domain {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticsearch-domain-snapshotoptions.html
     pub struct SnapshotOptions_ {
-        pub automated_snapshot_start_hour: Option<i64>,
+        pub automated_snapshot_start_hour: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -494,7 +494,7 @@ pub mod domain {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticsearch-domain-zoneawarenessconfig.html
     pub struct ZoneAwarenessConfig_ {
-        pub availability_zone_count: Option<i64>,
+        pub availability_zone_count: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
